test: Added cluster.Consolidated() checks to test
